Rm redundant lmd<->ffg check for fork choice attestation
